donatio-chatbot/pyproject.toml
[project]
name = "donatio-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.69",
    "langchain-groq>=0.3.6",
    "langgraph>=0.5.3",
    "numpy>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "typing-extensions>=4.14.0",
    "uvicorn>=0.35.0",
]
